trigger_dbt_function/requirements.txt
functions-framework==3.*
google-cloud-run==1.15.0